Revert "register api"
