+++ app/build.gradle
@@ -1,5 +1,4 @@
 apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
 
 android {
     compileSdkVersion 29
@@ -34,24 +33,5 @@ dependencies {
     implementation 'com.google.android.material:material:1.0.0'
     //noinspection GradleCompatible
     implementation 'com.android.support:design:28.0.0'
-// Retrofit
-    implementation 'com.squareup.retrofit2:retrofit:2.1.0'
 
-// JSON Parsing
-    implementation 'com.google.code.gson:gson:2.6.1'
-    implementation 'com.squareup.retrofit2:converter-gson:2.1.0'
-
-// recyclerview
-    implementation "androidx.recyclerview:recyclerview:1.1.0"
-    implementation "androidx.core:core-ktx:+"
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-
-//    RxJava
-    implementation 'io.reactivex.rxjava2:rxandroid:2.0.2'
-    implementation 'io.reactivex.rxjava2:rxjava:2.1.13'
-    implementation 'com.squareup.retrofit2:adapter-rxjava2:2.4.0'
-}
-repositories {
-    maven { url 'https://dl.bintray.com/kotlin/kotlin-eap' }
-    mavenCentral()
 }
