+++ app/src/main/java/com/example/meetup/InforLoginFragment.java
@@ -1,4 +1,4 @@
-package com.example.meetup.view;
+package com.example.meetup;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -10,8 +10,6 @@ import android.widget.TextView;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
-
-import com.example.meetup.R;
 
 public class InforLoginFragment extends Fragment implements View.OnClickListener {
     TextView tvForgotPassword;
