+++ app/src/main/java/com/example/meetup/InforSignupFragment.java
@@ -1,0 +1,25 @@
+package com.example.meetup;
+
+import android.os.Bundle;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+public class InforSignupFragment extends Fragment {
+    public InforSignupFragment() {
+        // Required empty public constructor
+    }
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        View view = inflater.inflate(R.layout.fragment_signup_infor, container, false);
+        return view;
+    }
+
+}
