+++ app/src/main/java/com/example/meetup/LoginFragment.java
@@ -1,4 +1,4 @@
-package com.example.meetup.view;
+package com.example.meetup;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -7,12 +7,11 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.ImageView;
+import android.widget.TextView;
 
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
-
-import com.example.meetup.R;
 
 public class LoginFragment extends Fragment {
     Button btnSignUp;
