--- app/src/main/java/com/example/meetup/model/User.java
@@ -1,44 +1,0 @@
-package com.example.meetup.model;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import retrofit2.Response;
-
-public class User {
-    public String email;
-    public String name;
-    public String password;
-
-    public User(String email, String name, String password) {
-        this.email = email;
-        this.name = name;
-        this.password = password;
-    }
-
-    public User(User response) {
-        this.token = token;
-    }
-
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    @SerializedName("status")
-    @Expose
-    private Integer status;
-    @SerializedName("token")
-    @Expose
-    private String token;
-
-
-}
