--- app/src/main/java/com/example/meetup/repositories/UserRepository.java
@@ -1,19 +1,0 @@
-package com.example.meetup.repositories;
-
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
-
-public class UserRepository {
-
-    private static Retrofit retrofit = null;
-
-    public static Retrofit getClient(String baseUrl) {
-        if (retrofit==null) {
-            retrofit = new Retrofit.Builder()
-                    .baseUrl(baseUrl)
-                    .addConverterFactory(GsonConverterFactory.create())
-                    .build();
-        }
-        return retrofit;
-    }
-}
