--- app/src/main/java/com/example/meetup/services/ApiUtils.java
@@ -1,11 +1,0 @@
-package com.example.meetup.services;
-
-import com.example.meetup.repositories.UserRepository;
-
-public class ApiUtils {
-    public static final String BASE_URL = "http://meetup.rikkei.org/api/v0/";
-
-    public static UserService getUserService() {
-        return UserRepository.getClient(BASE_URL).create(UserService.class);
-    }
-}
