+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
-    tools:context=".view.LoginActivity">
+    tools:context=".LoginActivity">
     <!--    Login & signup-->
     <ScrollView
         android:layout_width="match_parent"
