+++ app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
-    tools:context=".view.LoginActivity">
+    tools:context=".LoginActivity">
 
     <Button
         android:id="@+id/btnSignUp"
