+++ app/src/main/res/layout/fragment_login_infor.xml
@@ -126,7 +126,7 @@
                 android:layout_marginStart="16dp"
                 android:layout_marginTop="260dp"
                 android:layout_marginEnd="16dp"
-                android:background="@drawable/ic_btn_disable"
+                android:background="@drawable/ic_rectangle_btn"
                 android:fontFamily="sans-serif"
                 android:text="ĐĂNG NHẬP"
                 android:textColor="#FFFFFF"
